[JENKINS-30820] use classloader for stackMap references.
To generate stackFrames and for some parts of this the common ancestor of
2 classes is required.  ASM loads these classes using Class.forName()
using the classloader of the ASM classes.  In order to resolve this we
provide our own ClassWriter implementation which would be able to load
the classes using the plugins classloader.
Java classes can contain circular references so if multiple classes in
this circular chain require re-writing then this code will either
depadlock or enter an infinite loop (not sure which) but this would happen
less than the current situation which is to blow up with a
ClassNotFoundException.