[JENKINS-64479] do not use JUL by default for logging
Logging during classloading is a surefire way to deadlock.
but we still want to have logs and in some cases we also want those logs
to be in the same place as other logs.

so introduce a new helper that will either output to System.out
(default) or use JUL (opt in) at the risk of deadlocks