[JENKINS-64479] ensure all logging is async to avoid deadlocks
Logging during classloading is a surefire way to deadlock.
but we still want to have logs and in some cases we also want those logs
to be in the same place as other logs.

So introduce a new helper that punts the logging to a new Thread so that
we do not deadlock.